# Laravel/projecte_final/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example for response examples value",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/Projecte_Final/Laravel/projecte_final/public/index.php",
            "description": "L5 Swagger OpenApi Server"
        }
    ],
    "paths": {
...
        "/api/treballadors": {
            "get": {
                "tags": [
                    "Treballadors"
                ],
                "summary": "Veure tots els treballadors.",
                "operationId": "6f1a7dff10133be3b977a19fd9f9fc24",
                "responses": {
                    "200": {
                        "description": "Retorna tots els treballadors."
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/treballador/{id}": {
            "get": {
                "tags": [
                    "Treballadors"
                ],
                "summary": "Veure totes els treballadors.",
                "operationId": "eade75e03cdd064534ec5ed5b9b121b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter with mutliple examples",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna totes les treballadorss."
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Treballadors"
                ],
                "summary": "Update treballador",
                "operationId": "d04ef28447c52d6bc381d8f13d1dd849",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "treballador"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "treballador@gmail.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "treballador"
                    },
                    {
                        "name": "nom",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "treballador"
                    },
                    {
                        "name": "cognoms",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "treballador"
                    },
                    {
                        "name": "nif",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "12345"
                    },
                    {
                        "name": "sou",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "double"
                        },
                        "example": "1500.49"
                    },
                    {
                        "name": "carrec",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "String"
                        },
                        "example": "treballador"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Treballadors"
                ],
                "summary": "Delete treballador",
                "operationId": "fd2488c6da6337d7ac8cf7d47a373c20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Cita not found"
                    }
                }
            }
        },
        "/api/treballador": {
            "post": {
                "tags": [
                    "Treballadors"
                ],
                "summary": "Add a new treballador",
                "operationId": "c5c5b221d0dfe5cd8e6421ad6a77f906",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "treballador"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "treballador@gmail.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "treballador"
                    },
                    {
                        "name": "nom",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "treballador"
                    },
                    {
                        "name": "cognoms",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "treballador"
                    },
                    {
                        "name": "nif",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "12345"
                    },
                    {
                        "name": "sou",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "double"
                        },
                        "example": "1500.49"
                    },
                    {
                        "name": "carrec",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "String"
                        },
                        "example": "treballador"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}